move writeFile to utils
